feat: add Header comp to cyber-awareness and directory-privacy pages
