+++ components/privacy-directory/FAQSection.tsx
